\name{cjoint-package}
\alias{cjoint-package}
\alias{cjoint}
\docType{package}
\title{
cjoint: A Package for Estimating Average Marginal Component-specific Effects for Conjoint Survey Experiments
}
\description{
This package allows researchers to estimate the causal effects of attributes in conjoint survey experiments. It implements the Average Marginal Component-specific Effects (AMCE) estimator presented in Hainmueller, J., Hopkins, D., and Yamamoto T. (2014) Causal Inference in Conjoint Analysis: Understanding Multi-Dimensional Choices via Stated Preference Experiments. Political Analysis 22(1):1-30
}
\details{
\tabular{ll}{
Package: \tab cjoint\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2023-08-21\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Authors: Soubhik Barari, Elissa Berwick, Jens Hainmueller, Daniel Hopkins, Sean Liu, Anton Strezhnev, Teppei Yamamoto

Maintainer: Anton Strezhnev <astrezhnev@uchicago.edu>

}
\references{
Hainmueller, J., Hopkins, D., and Yamamoto T. (2014) Causal Inference in Conjoint Analysis: Understanding Multi-Dimensional Choices via Stated Preference Experiments. Political Analysis 22(1):1-30
}

\keyword{ package }

