% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean.names}
\alias{clean.names}
\title{clean.names}
\usage{
clean.names(str)
}
\arguments{
\item{str}{String to be cleaned.}
}
\description{
A function for removing ALL punctuation, symbols, and spaces from string
from elements of vector "vec"
}
