% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_se_glm.R
\name{cluster_se_glm}
\alias{cluster_se_glm}
\title{Clustered Standard Errors}
\usage{
cluster_se_glm(model, cluster)
}
\arguments{
\item{model}{Input model}

\item{cluster}{Cluster variable}
}
\description{
Clustered Standard Errors
}
