\name{conjoint_data}
\alias{conjoint_data}
\docType{data}
\title{
Immigration Conjoint Experiment Dataset from Hainmueller et. al. (2014)
}
\description{
A dataset containing the results of a conjoint survey of a representative sample of American adults who were asked to choose which hypothetical immigrants they think should be admitted into the United States. Each row corresponds to a single profile presented to the respondent.
}

\keyword{datasets}
