% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional_effects.R
\name{get.conditional.effects}
\alias{get.conditional.effects}
\title{Conditional Estimates}
\usage{
get.conditional.effects(
  object,
  conditional.levels,
  current.effect,
  current.level,
  mod.var
)
}
\arguments{
\item{object}{AMCE object}

\item{conditional.levels}{List containing levels 
at which conditional effects will be calculated}

\item{current.effect}{Name of current conditional variable}

\item{current.level}{Value of current level of current conditioning variable}

\item{mod.var}{Name of profile attribute being modified by 
conditioning variable}
}
\description{
Getting conditional estimates given
(1) AMCE object
(2) LIST containing levels at which conditional effects will be calculated
(3) NAME of current conditional variable
(4) VALUE of current level of current conditioning variable
(5) NAME of profile attribute being modified by conditioning variable
}
