% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_amce.R
\name{print.summary.amce}
\alias{print.summary.amce}
\title{Print Summary Function for Results of Main AMCE Function}
\usage{
\method{print}{summary.amce}(x, digits = 5, ...)
}
\description{
Print Summary Function for Results of Main AMCE Function
}
