% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{re.escape}
\alias{re.escape}
\title{Replace Values for Escape Expressions}
\usage{
re.escape(strings)
}
\arguments{
\item{strings}{Input strings}
}
\description{
Replace Values for Escape Expressions
}
