\name{view}
\alias{view}
\alias{view.amce}

\title{
Graphical interfaces for conjoint analysis
}
\description{
Launches a graphical interface (in R Shiny) for a given conjoint analysis operation. This allows the user to interactively perform a conjoint analysis task without having to specify any code in R. Graphical interface is launched locally in user's default web browser. It is recommended that this feature be used for exploratory analysis only.
}
\examples{
\dontrun{
# Launch graphical interface to interactively run AMCE
view(amce) 
}
}