% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R
\name{amce}
\alias{amce}
\title{Average Marginal Component Effects (AMCE) Calculation}
\usage{
amce(
  formula,
  data,
  design = "uniform",
  respondent.varying = NULL,
  subset = NULL,
  respondent.id = NULL,
  cluster = TRUE,
  na.ignore = FALSE,
  weights = NULL,
  baselines = NULL
)
}
\description{
Average Marginal Component Effects (AMCE) Calculation
}
