% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDesign.R
\name{makeDesign}
\alias{makeDesign}
\title{Make a Design Matrix}
\usage{
makeDesign(
  type = "file",
  J = NULL,
  filename = NULL,
  attribute.levels = NULL,
  constraints = NULL,
  level.probs = NULL,
  tol = 1e-14
)
}
\description{
Make a Design Matrix
}
