% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amce.R
\name{plot.amce}
\alias{plot.amce}
\title{Plot AMCE Estimates}
\usage{
\method{plot}{amce}(
  x,
  main = "",
  xlab = "Change in E[Y]",
  ci = 0.95,
  colors = NULL,
  xlim = NULL,
  breaks = NULL,
  labels = NULL,
  attribute.names = NULL,
  level.names = NULL,
  label.baseline = TRUE,
  text.size = 11,
  text.color = "black",
  point.size = 0.5,
  dodge.size = 0.9,
  plot.theme = NULL,
  plot.display = "all",
  facet.names = NULL,
  facet.levels = NULL,
  group.order = NULL,
  font.family = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "amce", a result of a call to amce}

\item{main}{Title of the plot.}

\item{xlab}{Label of the x-axis of the plot (AMCE or ACIE).
Default is "Change in E[Y]"}

\item{ci}{Levels for confidence intervals to plot around point estimates.
Must be between 0 and 1. Default is .95}

\item{colors}{Vector of color names to be used for points and
confidence intervals. The plot function will alternate between the colors
in the vector for each attribute being plotted.
If NULL, plot will use a default ggplot2 color scheme.}

\item{xlim}{Numeric vector denoting the upper and lower bounds of the
x-axis in the plot. If NULL the plot function will automatically set a
range that includes all effect estimates.}

\item{breaks}{Numeric vector denoting where x-axis tick marks should be
placed. If NULL plot will use ggplot2 defaults.}

\item{labels}{Vector denoting how x-axis tick marks should be labeled.
If NULL plot will use ggplot2 defaults.}

\item{attribute.names}{Character vector of attribute names to be plotted as
labels. By default plot.amce will use the attribute names in the "amce"
object passed to it.}

\item{level.names}{A list containing character vector elements with names
in attribute.names. Each character vector in the list contains the level
names to be plotted as labels beneath the corresponding attribute.
By default plot.amce will use the level names in the "amce" object
passed to it.}

\item{label.baseline}{If TRUE, the baseline levels for each attribute will
be labeled as such. Defaults to TRUE.}

\item{text.size}{Size of text. Defaults to 11.}

\item{text.color}{Color of text in plot. Defaults to "black".}

\item{point.size}{Size of points in the plot. Defaults to 0.5.}

\item{dodge.size}{Width to dodge overlaps to the side. Defaults to 0.9.}

\item{plot.theme}{A ggplot2 'theme' object to be added to the plot.
If NULL, defaults to black-and-white theme. Note that passing a theme
object will override text and point color/size options.}

\item{plot.display}{Character string, one of "all", "unconditional", or
"interaction". Option "all" will display both unconditional and interaction
estimates. The "unconditional" option will display only 1 plot for
unconditional estimates (both AMCE and ACIE) ignoring any facets provided
to "facet.names" or respondent-varying characteristics.
Option "interaction" will drop the unconditional plot and instead display
only (unconditional) ACIE's or estimates conditional on respondent-varying
characteristics as specified in the user-supplied option "facet.names".
Defaults to "all".}

\item{facet.names}{To facet plots (i.e., make separate plots for each
value of a variable) give "facet.names" a vector of character strings
containing the names of the variable(s) (either profile attribute or
respondent-varying) to facet by. Unless given specific levels in
"facet.levels", the plotted levels will consist of all levels of a factor
variable or the quantiles of a continuous variable. Multiple facet
variables cannot currently be varied at the same time within the same plot.
Instead conditional effects will be calculated for one facet at a time
while others are held at their first value in facet.levels, by default
the bottom quantile for continuous variables and the baseline for factors.}

\item{facet.levels}{To manually set facet levels, provide a list to
"facet.levels". Names of list entries should correspond with variable
names. The content of each entry should be a vector giving the desired
levels, whether factors or continuous. To change the displayed names of
the levels, assign names to each vector entry.}

\item{group.order}{To manually set the order of the attributes, provide
a vector to "group.order". Names of the vector entries should correspond
with name of the attribute.}

\item{font.family}{Will be passed to the ggplot function as the argument
for font.family. If NULL, defaults will be used.}

\item{...}{Other graphical parameters passed to ggplot.}
}
\description{
Plot AMCE Estimates
}
