% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amce.R
\name{plot.amce}
\alias{plot.amce}
\title{Plot AMCE function}
\usage{
\method{plot}{amce}(
  x,
  main = "",
  xlab = "Change in E[Y]",
  ci = 0.95,
  colors = NULL,
  xlim = NULL,
  breaks = NULL,
  labels = NULL,
  attribute.names = NULL,
  level.names = NULL,
  label.baseline = TRUE,
  text.size = 11,
  text.color = "black",
  point.size = 0.5,
  dodge.size = 0.9,
  plot.theme = NULL,
  plot.display = "all",
  facet.names = NULL,
  facet.levels = NULL,
  group.order = NULL,
  font.family = NULL,
  ...
)
}
\description{
Plot AMCE function
}
