% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_amce.R
\name{print.summary.amce}
\alias{print.summary.amce}
\title{Printing Summaries of AMCE Estimates}
\usage{
\method{print}{summary.amce}(x, digits = 5, ...)
}
\arguments{
\item{x}{An object of class "summary.amce", 
a result of a call to summary.amce.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments from other methods.}
}
\description{
Printing Summaries of AMCE Estimates
}
