% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read.with.qualtRics}
\alias{read.with.qualtRics}
\title{Read in Qualtrics Data (The Output from qualtRics Package)}
\usage{
read.with.qualtRics(
  filename,
  responses = NULL,
  covariates = NULL,
  respondentID = NULL,
  letter = "F",
  new.format = FALSE,
  ranks = NULL
)
}
\description{
Read in Qualtrics Data (The Output from qualtRics Package)
}
