% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_amce.R
\name{summary.amce}
\alias{summary.amce}
\title{Summary Function for Results of Main AMCE Function}
\usage{
\method{summary}{amce}(object, covariate.values = NULL, ...)
}
\description{
Summary Function for Results of Main AMCE Function
}
